graft-backend/requirements.txt
anthropic==0.52.2
fastapi==0.115.12
gitpython==3.1.43
httpx==0.28.1
networkx==3.4.2
pydantic==2.11.3
python-dotenv==1.0.1
tree-sitter==0.24.0
tree-sitter-languages==1.10.2
uvicorn==0.34.0
